multiple deployers: Ignore .debug files
This commit replicates a pattern from the TextToSpeechPluginsDeployer
across a range of other deployer classes.

Essentially - it's extremely unlikely that composing an AppImage
with debug symbols is what a user actually wants to do. I'm also
not aware of a good story for connecting debug symbols to an AppImage
at all.

As a result, let's make a sensible choice - and not package them.
Trust that developers will use their underlying build system for Debug,
and that release tracking will allow developers to tie an AppImage
back to a source control revision.

Finally, note that this works around an interesting aarch64 bug -
where the .debug files that are created are marked as x86_64 ELF
binaries - even when you are _building on an aarch64 host_.